Done with organization for the most part. Inserted all the routes in their respective files. Inserted all the modules in the models folder and the functions in the functions folder.
