+++ models/adminModel.js
@@ -9,7 +9,6 @@ const adminSchema = new mongoose.Schema(
 			required: true,
 			unique: true,
 			trim: true,
-			lowercase: true,
 		},
 		password: {
 			type: String,
