+++ models/studentModel.js
@@ -2,29 +2,33 @@ const mongoose = require("mongoose");
 const passportLocalMongoose = require("passport-local-mongoose");
 const passport = require("passport");
 
-const studentSchema = new mongoose.Schema({
-	username: {
-		type: String,
-		required: true,
-		unique: true,
-		trim: true,
-		lowercase: true,
+const studentSchema = new mongoose.Schema(
+	{
+		username: {
+			type: String,
+			required: true,
+			unique: true,
+			trim: true,
+		},
+		password: {
+			type: String,
+		},
+		joined: {
+			type: Date,
+			default: Date.now,
+		},
+		secret: {
+			type: String,
+		},
+		class: {
+			type: String,
+			default: "",
+		},
 	},
-	password: {
-		type: String,
-	},
-	joined: {
-		type: Date,
-		default: Date.now,
-	},
-	secret: {
-		type: String,
-	},
-	class: {
-		type: String,
-		default: "",
-	},
-});
+	{
+		timestamps: true, // Automatically adds `createdAt` and `updatedAt` fields
+	}
+);
 
 // studentSchema.plugin(passportLocalMongoose);
