+++ models/teacherModel.js
@@ -2,25 +2,26 @@ const mongoose = require("mongoose");
 const passportLocalMongoose = require("passport-local-mongoose");
 const passport = require("passport");
 
-const teacherSchema = new mongoose.Schema({
-	username: {
-		type: String,
-		required: true,
-		unique: true,
-		trim: true,
-		lowercase: true,
+const teacherSchema = new mongoose.Schema(
+	{
+		username: {
+			type: String,
+			required: true,
+			unique: true,
+			trim: true,
+		},
+		password: {
+			type: String,
+		},
+
+		secret: {
+			type: String,
+		},
 	},
-	password: {
-		type: String,
-	},
-	joined: {
-		type: Date,
-		default: Date.now,
-	},
-	secret: {
-		type: String,
-	},
-});
+	{
+		timestamps: true, // Automatically adds `createdAt` and `updatedAt` fields
+	}
+);
 
 teacherSchema.plugin(passportLocalMongoose);
 const Teacher = new mongoose.model("Teacher", teacherSchema);
