+++ views/classesControl.ejs
@@ -4,6 +4,7 @@
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<title>Classes Control Faith Baptist Academy</title>
+		<!-- TODO ADD FAVICON -->
 		<link rel="icon" href="img/favicon.ico" />
 		<!-- CSS FILES -->
 		<link
