+++ views/dashboardAdmin.ejs
@@ -4,6 +4,10 @@
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<title>Admin Portal Faith Baptist Academy</title>
+
+				<!-- TODO ADD FAVICON -->
+				<link rel="icon" href="img/favicon.ico" />
+
 		<!-- CSS FILES -->
 		<link rel="stylesheet" type="text/css" href="css/cards.css" />
 
@@ -19,6 +23,7 @@
 			href="https://cdn.jsdelivr.net/npm/uikit@latest/dist/css/uikit.min.css"
 		/>
 		<link rel="stylesheet" type="text/css" href="css/dashboard.css" />
+		
 	</head>
 	<body>
 		<!--HEADER-->
@@ -649,6 +654,105 @@
 					data-uk-grid
 					uk-sortable="handle: .sortable-icon"
 				>
+
+				<!-- panel -->
+				<div class="uk-width-1-2@l" id="classes">
+					<div class="uk-card">
+						<div class="uk-card-header">
+							<div class="uk-grid uk-grid-small">
+								<div class="uk-width-auto">
+									<h4>Classes</h4>
+								</div>
+								<div class="uk-width-expand uk-text-right panel-icons">
+									<a
+										href="#"
+										class="uk-icon-link sortable-icon"
+										title="Move"
+										data-uk-tooltip
+										data-uk-icon="icon: move"
+									></a>
+					
+									<a
+										href="#modal-center1"
+										class="uk-icon-link"
+										title="Add Class"
+										data-uk-tooltip
+										data-uk-toggle
+										
+									><span data-uk-icon="icon: plus" class="uk-text-success "></span></a>
+									<a
+										href="#"
+										class="uk-icon-link"
+										title="Close"
+										data-uk-tooltip
+										data-uk-icon="icon: close"
+									></a>
+								</div>
+							</div>
+						</div>
+						<div>
+							<div
+								class="uk-child-width-1-3@m uk-grid-small uk-grid-match"
+								uk-grid
+								id="homeClassCards"
+							>
+						<!-- card -->
+					
+
+				
+						
+
+							<div class="nature-card" data-tags="a beautiful landscape - nature outdoor">
+								<div class="uk-card uk-card-small uk-card-default">
+									<div class="uk-card-header">
+										<div class="uk-grid uk-grid-small uk-text-small" data-uk-grid>
+											<div class="uk-width-expand">
+												<span class="cat-txt"> Class Name   </span>
+											</div>
+											<div class="uk-width-auto uk-text-right uk-text-muted">
+												<span data-uk-icon="icon:clock; ratio: 0.8"></span> Stream(Soon!)
+											</div>
+										</div>
+									</div>
+									<div class="uk-card-body">
+										<h6 class="uk-margin-small-bottom uk-margin-remove-adjacent uk-text-bold">Class Teacher</h6>
+									</div>
+									<div class="uk-card-footer">
+										<div class="uk-grid uk-grid-small uk-grid-divider uk-flex uk-flex-middle" data-uk-grid>
+											<div class="uk-width-expand uk-text-small text-nowrap overflow-hidden" 
+											data-uk-tooltip
+											title="Class Members"
+											>
+												30 Members  
+											</div>
+											<div class="uk-width-auto uk-text-right">
+												<a href="#" data-uk-tooltip="title: Twitter" class="uk-icon-link" data-uk-icon="icon:twitter; ratio: 0.8"></a>
+												<a href="#" data-uk-tooltip="title: Instagram" class="uk-icon-link" data-uk-icon="icon:instagram; ratio: 0.8"></a>
+												<a href="#" data-uk-tooltip="title: Behance" class="uk-icon-link" data-uk-icon="icon:behance; ratio: 0.8"></a>
+												<a href="#" data-uk-tooltip="title: Pinterest" class="uk-icon-link" data-uk-icon="icon:pinterest; ratio: 0.8"></a>
+											</div>
+											<div class="uk-width-auto uk-text-right">
+												<form id="deleteForm" action="/classes/delete/" method="post">
+													<button
+														type="button"
+														class="uk-icon-link"
+														title="Remove Class"
+														data-uk-tooltip
+														data-uk-icon="icon: trash"
+														onclick="confirmDelete()"
+													></button>
+												</form>
+											</div>
+										</div>
+									</div>
+								</div>
+							</div>
+							<!--/ card-->
+							</div>
+						</div>
+					</div>
+				</div>
+				<!-- /panel -->
 					<!-- panel -->
 					<div class="uk-width-1-2@l">
 						<div class="uk-card uk-card-default uk-card-small uk-card-hover">
@@ -688,6 +792,7 @@
 						</div>
 					</div>
 					<!-- /panel -->
+					 
 
 					<!-- Table Added  -->
 
@@ -759,108 +864,7 @@
 					<!-- /panel -->
 
 					<!-- Added Actions  -->
-					<!-- panel -->
-					<div class="uk-width-1-2@l" id="classes">
-						<div class="uk-card">
-							<div class="uk-card-header">
-								<div class="uk-grid uk-grid-small">
-									<div class="uk-width-auto">
-										<h4>Classes</h4>
-									</div>
-									<div class="uk-width-expand uk-text-right panel-icons">
-										<a
-											href="#"
-											class="uk-icon-link sortable-icon"
-											title="Move"
-											data-uk-tooltip
-											data-uk-icon="icon: move"
-										></a>
-						
-										<a
-											href="#modal-center1"
-											class="uk-icon-link"
-											title="Add Class"
-											data-uk-tooltip
-											data-uk-toggle
-											
-										><span data-uk-icon="icon: plus" class="uk-text-success "></span></a>
-										<a
-											href="#"
-											class="uk-icon-link"
-											title="Close"
-											data-uk-tooltip
-											data-uk-icon="icon: close"
-										></a>
-									</div>
-								</div>
-							</div>
-							<div>
-								<div
-									class="uk-child-width-1-3@m uk-grid-small uk-grid-match"
-									uk-grid
-									id="homeClassCards"
-								>
-							<!-- card -->
-						
-
 					
-							
-
-								<div class="nature-card" data-tags="a beautiful landscape - nature outdoor">
-									<div class="uk-card uk-card-small uk-card-default">
-										<div class="uk-card-header">
-											<div class="uk-grid uk-grid-small uk-text-small" data-uk-grid>
-												<div class="uk-width-expand">
-													<span class="cat-txt"> Class Name   </span>
-												</div>
-												<div class="uk-width-auto uk-text-right uk-text-muted">
-													<span data-uk-icon="icon:clock; ratio: 0.8"></span> Stream(Soon!)
-												</div>
-											</div>
-										</div>
-										<div class="uk-card-body">
-											<h6 class="uk-margin-small-bottom uk-margin-remove-adjacent uk-text-bold">Class Teacher</h6>
-										</div>
-										<div class="uk-card-footer">
-											<div class="uk-grid uk-grid-small uk-grid-divider uk-flex uk-flex-middle" data-uk-grid>
-												<div class="uk-width-expand uk-text-small text-nowrap overflow-hidden" 
-												data-uk-tooltip
-												title="Class Members"
-												>
-													30 Members  
-												</div>
-												<div class="uk-width-auto uk-text-right">
-													<a href="#" data-uk-tooltip="title: Twitter" class="uk-icon-link" data-uk-icon="icon:twitter; ratio: 0.8"></a>
-													<a href="#" data-uk-tooltip="title: Instagram" class="uk-icon-link" data-uk-icon="icon:instagram; ratio: 0.8"></a>
-													<a href="#" data-uk-tooltip="title: Behance" class="uk-icon-link" data-uk-icon="icon:behance; ratio: 0.8"></a>
-													<a href="#" data-uk-tooltip="title: Pinterest" class="uk-icon-link" data-uk-icon="icon:pinterest; ratio: 0.8"></a>
-												</div>
-												<div class="uk-width-auto uk-text-right">
-													<form id="deleteForm" action="/classes/delete/" method="post">
-														<button
-															type="button"
-															class="uk-icon-link"
-															title="Remove Class"
-															data-uk-tooltip
-															data-uk-icon="icon: trash"
-															onclick="confirmDelete()"
-														></button>
-													</form>
-												</div>
-											</div>
-										</div>
-									</div>
-								</div>
-					
-						
-								<!--/ card-->
-				
-									
-								</div>
-							</div>
-						</div>
-					</div>
-					<!-- /panel -->
 
 					<!-- panel -->
 					<div class="uk-width-1-2@l">
