+++ views/dashboardStudent.ejs
@@ -21,6 +21,7 @@
 		/>
 		<link rel="stylesheet" type="text/css" href="css/dashboard.css" />
 	</head>
+
 	<body>
 		<!--HEADER-->
 		<header id="top-head" class="uk-position-fixed">
