+++ views/dashboardTeacher.ejs
@@ -5,6 +5,10 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<title>Teachers Portal Faith Baptist Academy</title>
 		<!-- CSS FILES -->
+		<!-- TODO ADD FAVICON -->
+		<link rel="icon" href="img/favicon.ico" />
+
+
 		<link
 			href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
 			rel="stylesheet"
@@ -20,6 +24,7 @@
 		/>
 		<link rel="stylesheet" type="text/css" href="css/dashboard.css" />
 	</head>
+
 	<body>
 		<!--HEADER-->
 		<header id="top-head" class="uk-position-fixed">
