--- views/register1.ejs
@@ -1,196 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
-		<title>Faith Baptist Academy Register</title>
-		<link rel="icon" href="img/favicon.ico" />
-		<!-- CSS FILES -->
-		<link
-			rel="stylesheet"
-			type="text/css"
-			href="https://cdn.jsdelivr.net/npm/uikit@latest/dist/css/uikit.min.css"
-		/>
-	</head>
-	<body
-		class="uk-flex uk-flex-center uk-flex-middle uk-background-muted uk-height-viewport"
-		data-uk-height-viewport
-	>
-		<div
-			class="uk-position-bottom-center uk-position-small uk-visible@m uk-position-z-index"
-		>
-			<!-- TODO ADD COMPANY NAME  -->
-			<span class="uk-text-small uk-text-muted"
-				>© 2024 Company Name -
-				<a href="https://github.com/zzseba78/Kick-Off"
-					>Created by Dev Austine Alex &amp; Dev Elvis Kemoi</a
-				>
-				| Built with
-				<a
-					href="http://getuikit.com"
-					title="Visit UIkit 3 site"
-					target="_blank"
-					data-uk-tooltip
-					><span data-uk-icon="uikit"></span></a
-			></span>
-		</div>
-		<div class="uk-width-medium uk-padding-small">
-			<div class="uk-text-center uk-margin">
-				<!-- <img src="img/login-logo.svg" alt="Logo" /> -->
-				Faith Baptist Academy
-			</div>
-			<!-- REGISTER -->
-			<form class="toggle-class" action="/register" method="post">
-				<fieldset class="uk-fieldset">
-					<div class="uk-inline uk-width-1-1">
-						<h2>Register as ?</h2>
-						<span
-							class="uk-form-icon uk-form-icon-flip"
-							data-uk-icon="icon: user"
-						></span>
-					</div>
-					<div class="uk-margin-small">
-						<div class="form-check">
-							<input
-								class="form-check-input"
-								type="radio"
-								name="userType"
-								id="flexRadioDefault2"
-								value="Student"
-								checked
-							/>
-							<label class="form-check-label" for="flexRadioDefault2">
-								Student
-							</label>
-						</div>
-						<!-- <div class="form-check">
-							<input
-								class="form-check-input"
-								type="radio"
-								value="Teacher"
-								name="userType"
-								id="flexRadioDefault3"
-							/>
-							<label class="form-check-label" for="flexRadioDefault3">
-								Teacher
-							</label>
-						</div> -->
-						<div class="form-check">
-							<input
-								class="form-check-input"
-								type="radio"
-								value="Admin"
-								name="userType"
-								id="flexRadioDefault1"
-							/>
-							<label class="form-check-label" for="flexRadioDefault1">
-								Admin
-							</label>
-						</div>
-					</div>
-					<div class="uk-margin-small">
-						<div class="uk-inline uk-width-1-1">
-							<span
-								class="uk-form-icon uk-form-icon-flip"
-								data-uk-icon="icon: user"
-							></span>
-							<input
-								class="uk-input uk-border-pill"
-								required
-								placeholder="Username"
-								type="text"
-								name="username"
-							/>
-						</div>
-					</div>
-					<div class="uk-margin-small">
-						<div class="uk-inline uk-width-1-1">
-							<span
-								class="uk-form-icon uk-form-icon-flip"
-								data-uk-icon="icon: lock"
-							></span>
-							<input
-								class="uk-input uk-border-pill"
-								required
-								placeholder="Password"
-								type="password"
-								name="password"
-							/>
-						</div>
-					</div>
-
-					<div class="uk-margin-small">
-						<label
-							><input class="uk-checkbox" type="checkbox" /> Keep me logged
-							in</label
-						>
-					</div>
-					<div class="uk-margin-bottom">
-						<button
-							type="submit"
-							class="uk-button uk-button-primary uk-border-pill uk-width-1-1"
-						>
-							Register
-						</button>
-					</div>
-				</fieldset>
-				<div>
-					<p>
-						Already have an account?
-						<a href="/login" class="text-primary">Login</a>
-					</p>
-				</div>
-			</form>
-			<!-- /login -->
-
-			<!-- recover password -->
-			<!-- <form class="toggle-class" action="login-dark.html" hidden>
-				<div class="uk-margin-small">
-					<div class="uk-inline uk-width-1-1">
-						<span
-							class="uk-form-icon uk-form-icon-flip"
-							data-uk-icon="icon: mail"
-						></span>
-						<input
-							class="uk-input uk-border-pill"
-							placeholder="E-mail"
-							required
-							type="text"
-						/>
-					</div>
-				</div>
-				<div class="uk-margin-bottom">
-					<button
-						type="submit"
-						class="uk-button uk-button-primary uk-border-pill uk-width-1-1"
-					>
-						SEND PASSWORD
-					</button>
-				</div>
-			</form> -->
-			<!-- /recover password -->
-
-			<!-- action buttons -->
-			<!-- <div>
-				<div class="uk-text-center">
-					<a
-						class="uk-link-reset uk-text-small toggle-class"
-						data-uk-toggle="target: .toggle-class ;animation: uk-animation-fade"
-						>Forgot your password?</a
-					>
-					<a
-						class="uk-link-reset uk-text-small toggle-class"
-						data-uk-toggle="target: .toggle-class ;animation: uk-animation-fade"
-						hidden
-						><span data-uk-icon="arrow-left"></span> Back to Login</a
-					>
-				</div>
-			</div> -->
-			<!-- action buttons -->
-		</div>
-
-		<!-- JS FILES -->
-		<script src="https://cdn.jsdelivr.net/npm/uikit@latest/dist/js/uikit.min.js"></script>
-		<script src="https://cdn.jsdelivr.net/npm/uikit@latest/dist/js/uikit-icons.min.js"></script>
-	</body>
-</html>
