--- views/secrets.ejs
@@ -1,17 +1,0 @@
-<%- include('partials/header') %>
-
-<div class="jumbotron text-center">
-	<div class="container">
-		<i class="fas fa-key fa-6x"></i>
-		<h1 class="display-3">You've Discovered My Secret!</h1>
-
-		<hr />
-
-		<a class="btn btn-light btn-lg" href="/logout" role="button">Log Out</a>
-		<a class="btn btn-dark btn-lg" href="/submit" role="button"
-			>Submit a Secret</a
-		>
-	</div>
-</div>
-
-<%- include('partials/footer') %>
