+++ views/teachersControl.ejs
@@ -4,14 +4,16 @@
 		<meta charset="utf-8" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<title>Teachers View Faith Baptist Academy</title>
+		<!-- TODO ADD FAVICON -->
 		<link rel="icon" href="img/favicon.ico" />
-		<!-- CSS FILES -->
 		<link
 			href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css"
 			rel="stylesheet"
 			integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH"
 			crossorigin="anonymous"
 		/>
+		<link rel="stylesheet" type="text/css" href="css/cards.css" />
+
 		<link
 			rel="stylesheet"
 			type="text/css"
